pyproject.toml
[project]
name = "stage-1"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "aiohttp>=3.13.5",
    "django>=6.0.4",
    "django-cors-headers>=4.9.0",
    "django-ninja>=1.6.2",
    "django-ninja-jwt>=5.4.4",
    "djangorestframework>=3.17.1",
    "dotenv>=0.9.9",
    "httpx>=0.28.1",
    "psycopg[binary]>=3.3.3",
    "reportlab>=4.5.0",
    "requests>=2.33.1",
    "spire-pdf>=12.4.1",
]
